requirements.txt
# --- Web ve API ---
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# --- Firebase ---
firebase-admin==6.5.0
google-cloud-firestore==2.21.0      # ← düzeltildi
google-cloud-storage==2.16.0

# --- Veri modelleri & Ayarlar ---
pydantic==2.7.3
pydantic-settings==2.2.1

# --- Ödeme & Kargo ---
iyzipay==1.0.45
requests==2.32.3

# --- Background / Scheduler ---
APScheduler==3.10.4
tzdata==2024.1

# --- Loglama & Rate-limit ---
structlog==24.1.0
slowapi==0.1.9

# --- Test ---
pytest==8.2.0
pytest-asyncio==0.23.7
httpx==0.27.0

# --- Tip kontrolleri (opsiyonel) ---
mypy==1.10.0
types-requests==2.32.0.20240523

python-jose[cryptography]  # JWT üretip doğrulamak için

SQLAlchemy>=2.0
